+++ src/main/java/io/github/astrapi69/comparator/factory/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * This package provides a factory class for comparators
+ * <p>
+ * </p>
+ *
+ * @author Asterios Raptis
+ */
+package io.github.astrapi69.comparator.factory;