+++ src/main/java/io/github/astrapi69/comparator/object/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * This package provides comparators for Date, Locale, String and with the hashcode of the object
+ * <p>
+ * </p>
+ *
+ * @author Asterios Raptis
+ */
+package io.github.astrapi69.comparator.object;