+++ src/main/java/io/github/astrapi69/comparator/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * This package provides utility classes for comparators and set the sort order
+ * <p>
+ * </p>
+ *
+ * @author Asterios Raptis
+ */
+package io.github.astrapi69.comparator;